+++ server/model/garage.go
@@ -6,15 +6,15 @@ type Garage struct {
 	Latitude      float64 `json:"latitude,omitempty"`
 	Longituted    float64 `json:"longitute,omitempty"`
 	Level         int64   `json:"level,omitempty"`         //level reuired to unlock the garage
-	CoinsRequired int     `json:"coinsRequired,omitempty"` //coins required to unlock the garage
+	CoinsRequired int64   `json:"coinsRequired,omitempty"` //coins required to unlock the garage
 	Locked        bool    `json:"locked,omitempty"`
 }
 
 type GarageUpgrades struct {
 	GarageId      string `json:"garageId,omitempty"`
-	UpgradeLevel  int    `json:"upgradeLevel,omitempty"`  // level of the garage
+	UpgradeLevel  int64  `json:"upgradeLevel,omitempty"`  // level of the garage
 	UpgradeAmount int64  `json:"upgradeAmount,omitempty"` // amount required for the grage to be upgraded
-	CarLimit      int    `json:"carLimit,omitempty"`      //limit of cars a player can store in that garage
+	CarLimit      int64    `json:"carLimit,omitempty"`      //limit of cars a player can store in that garage
 }
 type GarageCarList struct {
 	PlayerId string `json:"playerId,omitempty"`
